feat: 36 refactor validator structure
