+++ src/validators/api/email-address.ts
@@ -1,5 +1,6 @@
 import { z } from 'zod';
-import { zMapAppToDbData, zMapDbToAppData } from './util';
+import { zMapDbToAppData } from './util/z-map-db-to-app-data';
+import { zMapAppToDbData } from './util/z-map-app-to-db-data';
 
 export const zEmailAddressDto = z.object({ email: z.string().min(1) });
