+++ src/validators/app/util/index.ts
@@ -1,0 +1,1 @@
+export * from './get-form-validator';
